fix du bug BDD qui plante avec l'activité NoteActivity

+++ app/src/main/java/com/thiti/blocnote/Model/DAO/DAOBase.java
@@ -40,7 +40,7 @@ public abstract class DAOBase<T> {
 
     public abstract void upgrade(T objet);
 
-    public abstract T find(int id);
+    public abstract T find(long id);
 
     public abstract List<T> all();
